add pragments
